[Perf] Enable low-precision Wan VAE execution

+++ vllm_omni/diffusion/models/wan2_2/pipeline_wan2_2.py
@@ -315,7 +315,7 @@ class Wan22Pipeline(nn.Module, CFGParallelMixin, ProgressBarMixin, DiffusionPipe
         ).to(self.device)
         attach_t5_encoder_hsdp_shard_conditions(self.text_encoder)
         self.vae = DistributedAutoencoderKLWan.from_pretrained(
-            model, subfolder="vae", torch_dtype=torch.float32, local_files_only=local_files_only
+            model, subfolder="vae", torch_dtype=dtype, local_files_only=local_files_only
         ).to(self.device)
 
         # Initialize transformers with correct config (weights loaded via load_weights)
